Add cleanupComparisonImages option in a custom way
This option does not use the native cleanupComparisonImages phantomCSS option, instead enable or disable the cleanup after the run.

The reason of this is because the feature implemented on phantomCSS 0.7.2 change the result path adding random subpaths to the final image directory.

+++ tasks/phantomcss.js
@@ -20,6 +20,7 @@ module.exports = function(grunt) {
 
         var options = this.options({
             addLabelToFailedImage: true,
+            cleanupComparisonImages: true,
             screenshots: 'screenshots',
             results: 'results',
             viewportSize: [1280, 800],
@@ -59,8 +60,9 @@ module.exports = function(grunt) {
             // Remove temporary file
             tempFile.unlink();
 
-
-            deleteDiffScreenshots();
+            if (options.cleanupComparisonImages) {
+              deleteDiffScreenshots();
+            }
 
             done(error || failureCount === 0);
         };
